+++ src/features/__init__.py
@@ -1,2 +1,3 @@
 from . import clearsky
+from . import transformers
 from . import preprocessing
